Rearrange armable callbacks
The armable had callbacks named for what they did (activate, arm,
etc.) rather than the event they handled.  We used a pattern in the
text field of methods to do the actual work, and callbacks which
called the needed method; we'll use that pattern for the armable.

We add simple activate(), deactivate(), arm(), and disarm() action
methods, and rename the actual event callbacks for the events they
handle.  The callbacks now call the new action methods.

The changes cascade into the toggle, but they are minor.