Add get/set for the composite focused child
We've had a get/set for the radio-related children, and that'll be
a great entry point for focus-related children too.

We'll take this opportunity to rearrange with a more common pattern
used in other widgets in the hierarchy - the top-level get/set call
a mid-level get/set_child, which then call out to the leaf
get/set_focused_child methods.  This also allows us to link in the
radio child methods in that same call tree.

The upshot is that we can perform set()s of the focused child from
the child itself.  There are some items which will set the focus,
but which we do not want to retain focus.  We also want "stop being
focused" to bubble up to the top-level in those cases, so a chained
callback will do nicely.