Use new focus-setting in composite to percolate up
Now that we can tell composite widgets what to focus, we can force
a given widget to be focused, or be able to drop focus from any
widgets, all the way to the top of the hierarchy.  The existing
mechanism already handles the top-down motion of focus, but the
upward motion is new.

Furthermore, the mouse button callback handler in the composite now
only sets focus on a button-down event.  Button-up events are not
particularly interesting as far as focus goes.  We'll also drop the
secondary "make sure the widget still exists" clause at the end of
that callback; as long as the event has been propagated, it's not
especially important if the widget is still around.  In future
commits, we'll probably have widgets use the new percolate-up to
drop focus as they become invisible or are deleted.

Re:  issue #4