Add focus handling for armable widgets
The armable has a little different relationship with "focus" as far
as keyboard action goes.  Once an armable is clicked, it's a strong
likelihood that we don't want to continue having focus, so we can
bubble that back up as we become disarmed.

Other widget sets also have visual indicators on keyboard focus for
buttons.  Often, it's a separate indication from the regular mouse-
enter appearance change.  For the time being, we'll keep things
more simple, and use the same activate action for mouse-enter and
focus-enter.  We do need to keep track of how many events have
activated the widget though, so we'll add an activations member.

Re:  issue #4