Convert composite focused child member to iterator
Once we start adding in the keyboard handling for moving focus, we
want to be able to move to the next or previous child.  Rather than
searching EVERY TIME, we'll keep an iterator as the focused child.
That way, when we move to the next, we increment the iterator, and
moving to the previous is a decrement.

In testing, we found that the set_focused_child() method was
overrunning the stack with an infinite recursion, because we
weren't checking for NULL -> end-of-list.  We'll make the initial
condition into a bailout clause which will allow us to handle
the rest of the logic more directly.

Re:  issue #4