Add tab handling for moving focus in composite
The main part of this whole exercise:  the ability to move focus
with the tab/shift-tab keystrokes.  The key handler gets an
additional clause at the beginning of the method, and we add
focus_next_child() and focus_previous_child() to do the actual
work.

We'll also update the test to exercise it a little.

Currently this is only handled by the top-level context, because
we've got no logic for descending into other composite widgets.

Re:  issue #4