Allow key focus to be owned by a child composite
The tab handling for keyboard focus has been only handled by the
context up to now, because we included no logic to cascade focus-
handling down into child composites.  The logic to handle this
is surprisingly simple; we need to move the "is the focused child
a composite" test into both the tab handler and the non-tab
handler.  As a result, we can move the dynamic cast to determine
whether our focused child is a composite outside of all our test
clauses.

While we're working on this, we noticed a case in focus_next_child
in which we could end up with a seg-fault, by trying to increment
an iterator pointing to the end of our children list.

We also augment the number of tabs and shift-tabs we send in our
test so that we wrap around the end of a composite child list both
ways.

Re:  issue #4