Add keyboard handling by armable widgets
Most other widget sets seem to use the space bar to activate
various button widgets.  We'll add some key callbacks to the
armable widget which call through the corresponding mouse button
callback lists; one for button/key down, one for button/key up.

We'll add some testing with the uitest program.

We added some logic a while ago to drop focus when a button-press
action was complete, but in retrospect, it seems like the wrong way
to go.  We'll drop that while we're here.

Re:  issue #4