Add shader validation
We've recently become aware of the glslangValidator program, and
figured it would be a nice option to have available, if it's
installed.

We split out the perl Test::More validation check within the
configure script, so we can independently enable the GUI tests or
the shader validation.  We also add a check for glslangValidator
itself.

We add a test file to go over all our shaders and see if the
validator finds them valid.