Move string conversion functions into their own file
There's no reason that, if we just want the UTF8-to-32-bit-string
and 32-bit-string-to-UTF8 functions, that we should need to include
the entire font rendering engine.  So we'll move the functions into
their own pair of files.  Only one widget file and a couple of test
files need to be updated for this move.