Initialize glm::mat4 objects correctly
Newer versions of GLM change the way they initialize their matrices
in that they just zero them out completely.  The proper way to
initialize them to the identity matrix (which is generally what we
want) is to give them a single 1.0f argument.  This is supposed to
be backward compatible with older versions of GLM, so we shouldn't
be causing any trouble for the zero people who use this toolkit.

Fortunately, we only have a few instances of glm::mat4, so this
change is small and relatively painless.

Ref:  https://github.com/g-truc/glm/issues/809