Clean up ui_defs.h includes
In trying to include only active.h within an r9client file, we found
that we had to include ui_defs.h first to pull in the GET_VA and
SET_VA macros.  The rect.h file uses those two macros, and is the base
for everything else in the widget set, so we should be able to resolve
the problem once and for all, in that single place.

We've included ui_defs.h directly in the bulk of the .cc files
throughout the widget set, and since rect.h is included everywhere, we
can drop those redundant includes.  It might speed up building by a
microscopic bit?  A couple of the test files also get to drop their
extraneous includes.

Re:  issue #89