Decouple key from character in text field
We assume that we will only ever get a key code or a character in any
key callback, but that's not necessarily a valid assumption.  Our GLFW
connector currently enforces this, but it may not be valid on any
other system (or in the future for the GLFW connector).

If we get a character, we'll insert it into the text field's string,
but otherwise, we'll treat it as whatever key it is.  It's not
important that there *only* be a character code, so we'll drop the
requirement that there *not* be a key code.