Return key codes unless doing text input
We may want to be able to use a character-producing key in "just a
key" context.  The way we do things now, with either a key code or a
character, and never both, is that we would have to have a bunch of
conditions, and the code to handle it would be extremely ugly.

We tried a character-to-key map, but the ultimate end of that is that
the map is the size of every glyph in the world.  About as far from
good as possible.  A better solution is to insert the character
callback only when a text field is active and remove it when the text
field loses keyboard focus.  We add a hook function pointer into the
text field class, and have our GLFW connector populate it as it starts
up.  In the connector, the hook function adds the char callback on
gaining focus and removes it on losing focus.

I'm not sure what other GL windowing environments might need, so this
may not be the best long-term solution.