Move to Github actions for CI
TravisCI costs too much, and Codecov doesn't do a very good job, so
we'll remove them in favor of Github actions.

We implement the build/test process as a reusable action, because it
will be useful in the push context that we implement here, but also in
a more extended context we'll add for pull-requests, which will add
coverage calculations.  We'll add that in a future commit.