Make synopses' variable names match the class
Most of the manual pages use the first character of the class in
question for the variable name in the synopsis, e.g. class `ui::foo`
would use `f`, and class `ui::bar` would use `b`.  Our template uses
`t`, and in a couple of the newer pages, we neglected to switch from
that to the appropriate letter.