Enable coverage summaries
If we have lcov and genhtml installed, we'll try to use them for
generating coverage summaries.

We won't even look for them if we can't or won't run the GUI test -
most of the toolkit has no coverage except for the GUI test, so the
summary would be pretty sparse without it.