Add radio box adapter for composite objects
This ended up rather simpler than expected.  The correct places for
the radio elements is within the enhanced composite, and the correct
place to trigger the behaviour is from the child toggle widgets -
they're the ones that accept the click action.

The child callback function gets added to all children regardless of
type - we were seeing strangeness while `dynamic_cast`ing within the
`add_child` method.  The callback function itself verifies that the
widget is of the correct type before it will try to set anything.  Not
ideal, but not a whole lot of overhead for non-toggle children.

This is currently a parallel implementation, and would likely not even
compile if used.  Swapping this in should be almost entirely deletions
from the composite.

Ref:  #94