Clean up dates in source file comment blocks
Having the last updated time/date stamp in every file causes way more
trouble than it's worth.  Reverts of commits become more painful,
merges become more painful, and git keeps track of all that stuff
anyway.

We also update the copyright years, turning them into a range of years
that the file was changed.  We're not counting this commit, because
nothing of interest is actually changing.